rename files and dirs
