+++ src/co/edu/uptc/models/Model202128687/Node.java
@@ -1,4 +1,4 @@
-package co.edu.uptc.models.Alex202128687;
+package co.edu.uptc.models.Model202128687;
 
 // Vamos a usar arboles binarios de busqueda
 public class Node<T>{
